requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
pydantic==2.3.0
pydantic-settings==2.0.3
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
spotipy==2.23.0
praw==7.7.1
google-api-python-client==2.97.0
apscheduler==3.10.4
redis==4.6.0
alembic==1.12.0
aiohttp==3.8.5
